Completed part 2 of day 11
Easier than expected just had to change the rate a which each distance increased.
But I was having an off by 1 error because when swapping 1 with 1000000 you take
out the 1 and add the million, but the way I was solving the problem I was
leaving the 1 in and adding a million on top of that making it 1000001, so
instead I should add 999999 because the 1 that makes it a million was left in.